+++ google-maps/ios/Plugin/CapacitorGoogleMapsPlugin.m
@@ -13,6 +13,7 @@ CAP_PLUGIN(CapacitorGoogleMapsPlugin, "CapacitorGoogleMaps",
    CAP_PLUGIN_METHOD(disableClustering, CAPPluginReturnPromise);
    CAP_PLUGIN_METHOD(destroy, CAPPluginReturnPromise);
    CAP_PLUGIN_METHOD(setCamera, CAPPluginReturnPromise);
+   CAP_PLUGIN_METHOD(getMapType, CAPPluginReturnPromise);
    CAP_PLUGIN_METHOD(setMapType, CAPPluginReturnPromise);
    CAP_PLUGIN_METHOD(enableIndoorMaps, CAPPluginReturnPromise);
    CAP_PLUGIN_METHOD(enableTrafficLayer, CAPPluginReturnPromise);
@@ -21,4 +22,6 @@ CAP_PLUGIN(CapacitorGoogleMapsPlugin, "CapacitorGoogleMaps",
    CAP_PLUGIN_METHOD(setPadding, CAPPluginReturnPromise);           
    CAP_PLUGIN_METHOD(onScroll, CAPPluginReturnPromise);
    CAP_PLUGIN_METHOD(getMapBounds, CAPPluginReturnPromise);
+   CAP_PLUGIN_METHOD(mapBoundsContains, CAPPluginReturnPromise);
+   CAP_PLUGIN_METHOD(mapBoundsExtend, CAPPluginReturnPromise);
 )