+++ share/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.0.0-alpha.1](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/share@4.1.1...@capacitor/share@5.0.0-alpha.1) (2023-03-16)
+
+
+### Features
+
+* **android:** Removing enableJetifier ([d66f9cb](https://github.com/ionic-team/capacitor-plugins/commit/d66f9cbd9da7e3b1d8c64ca6a5b45156867d4a04))
+
+
+
+
+
+## [4.1.1](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/share@4.1.0...@capacitor/share@4.1.1) (2023-02-03)
+
+
+### Bug Fixes
+
+* **share:** not sharing files on Android < 10 ([#1386](https://github.com/ionic-team/capacitor-plugins/issues/1386)) ([13673f5](https://github.com/ionic-team/capacitor-plugins/commit/13673f599a37f0553baa5b8e533fd7f163d9de69))
+
+
+
+
+
 # [4.1.0](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/share@1.1.2...@capacitor/share@4.1.0) (2022-11-16)